+++ derive/src/attribute/mod.rs
@@ -1,5 +1,7 @@
 pub mod field;
 pub mod item;
+pub mod message;
 pub mod repr;
+pub mod variant;
 
 pub use repr::Repr;