+++ src/classified.rs
@@ -1,9 +1,8 @@
-use crate::{Diagnostic, DiagnosticSeverity, code, traits::DiagnosticMessageResolver};
+use crate::{Diagnostic, DiagnosticSeverity, code};
 
 /// A [`Diagnostic`] with an explicit [severity](DiagnosticSeverity).
 ///
 /// This is constructed through the [`Diagnostic::classify`] method.
-#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct ClassifiedDiagnostic<S, D = code::DefaultDiscriminant> {
     /// The diagnostic being classified.
     pub(super) inner: Diagnostic<S, D>,
@@ -34,14 +33,12 @@ impl<S, D> ClassifiedDiagnostic<S, D> {
     /// Reports this diagnostic using the given configuration.
     ///
     /// See the [`ariadne`] documentation for more details.
-    pub fn report<Resolver, C>(self, config: ariadne::Config, cache: C) -> std::io::Result<()>
+    pub fn report<C>(self, config: ariadne::Config, cache: C) -> std::io::Result<()>
     where
         S: ariadne::Span,
         D: code::Discriminant,
-        Resolver: DiagnosticMessageResolver<D>,
         C: ariadne::Cache<S::SourceId>,
     {
-        self.inner
-            .report_with::<Resolver, _>(self.severity, config, cache)
+        self.inner.report_with(self.severity, config, cache)
     }
 }